Extract isolate harness/globals from string constants into real JS files
Move HARNESS_CODE and GLOBALS_MODULE from inline template literals in
index.ts into separate *.isolate.js files under src/isolate/. Wrangler's
rules config treats the *.isolate.js glob as Text modules, so they're
bundled as string constants at build time while being authored as proper
JS files with syntax highlighting and linting.
