+++ executor/packages/sandbox-host/src/index.ts
@@ -35,6 +35,15 @@
 
 import { WorkerEntrypoint } from "cloudflare:workers";
 
+// Import isolate modules as raw text — these are loaded as JS modules inside
+// the dynamic isolate, NOT executed in the host worker. The *.isolate.js
+// extension is mapped to Text type in wrangler.jsonc rules, so wrangler
+// bundles them as string constants instead of trying to execute them.
+// @ts-expect-error — wrangler Text module import (no TS declarations)
+import GLOBALS_MODULE from "./isolate/globals.isolate.js";
+// @ts-expect-error — wrangler Text module import (no TS declarations)
+import HARNESS_CODE from "./isolate/harness.isolate.js";
+
 // ── Types ────────────────────────────────────────────────────────────────────
 
 interface Env {
@@ -182,104 +191,10 @@ export class ToolBridge extends WorkerEntrypoint<Env> {
 // isolate. User code lives in a **separate** module (`user-code.js`) and is
 // imported by the harness. This prevents user code from accessing or
 // manipulating the harness's fetch handler, `req`, `env`, `ctx`, or `Response`.
-
-const GLOBALS_MODULE = `
-// Captured before user code module is evaluated (this module is imported first).
-export const ResponseJson = Response.json.bind(Response);
-`;
-
-const HARNESS_CODE = `
-import { ResponseJson as _ResponseJson } from "./globals.js";
-import { run } from "./user-code.js";
-
-const APPROVAL_DENIED_PREFIX = "APPROVAL_DENIED:";
-
-function formatArgs(args) {
-  return args.map((v) => {
-    if (typeof v === "string") return v;
-    try { return JSON.stringify(v); }
-    catch { return String(v); }
-  }).join(" ");
-}
-
-function createToolsProxy(bridge, path = []) {
-  const callable = () => {};
-  return new Proxy(callable, {
-    get(_target, prop) {
-      if (prop === "then") return undefined;
-      if (typeof prop !== "string") return undefined;
-      return createToolsProxy(bridge, [...path, prop]);
-    },
-    async apply(_target, _thisArg, args) {
-      const toolPath = path.join(".");
-      if (!toolPath) throw new Error("Tool path missing");
-      const input = args.length > 0 ? args[0] : {};
-      const result = await bridge.callTool(toolPath, input);
-      if (result.ok) return result.value;
-      if (result.denied) throw new Error(APPROVAL_DENIED_PREFIX + result.error);
-      throw new Error(result.error);
-    },
-  });
-}
-
-export default {
-  async fetch(req, env, ctx) {
-    const stdoutLines = [];
-    const stderrLines = [];
-
-    const appendStdout = (line) => {
-      stdoutLines.push(line);
-      ctx.waitUntil(env.TOOL_BRIDGE.emitOutput("stdout", line));
-    };
-    const appendStderr = (line) => {
-      stderrLines.push(line);
-      ctx.waitUntil(env.TOOL_BRIDGE.emitOutput("stderr", line));
-    };
-
-    const tools = createToolsProxy(env.TOOL_BRIDGE);
-    const console = {
-      log: (...args) => appendStdout(formatArgs(args)),
-      info: (...args) => appendStdout(formatArgs(args)),
-      warn: (...args) => appendStderr(formatArgs(args)),
-      error: (...args) => appendStderr(formatArgs(args)),
-    };
-
-    try {
-      const value = await run(tools, console);
-
-      if (value !== undefined) {
-        appendStdout("result: " + formatArgs([value]));
-      }
-
-      return _ResponseJson({
-        status: "completed",
-        stdout: stdoutLines.join("\\n"),
-        stderr: stderrLines.join("\\n"),
-        exitCode: 0,
-      });
-    } catch (error) {
-      const message = error instanceof Error ? error.message : String(error);
-      if (message.startsWith(APPROVAL_DENIED_PREFIX)) {
-        const denied = message.replace(APPROVAL_DENIED_PREFIX, "").trim();
-        appendStderr(denied);
-        return _ResponseJson({
-          status: "denied",
-          stdout: stdoutLines.join("\\n"),
-          stderr: stderrLines.join("\\n"),
-          error: denied,
-        });
-      }
-      appendStderr(message);
-      return _ResponseJson({
-        status: "failed",
-        stdout: stdoutLines.join("\\n"),
-        stderr: stderrLines.join("\\n"),
-        error: message,
-      });
-    }
-  },
-};
-`;
+//
+// Both HARNESS_CODE and GLOBALS_MODULE are imported as raw text from
+// `./isolate/harness.js` and `./isolate/globals.js` respectively, so they
+// can be authored as real JS files with proper syntax highlighting and linting.
 
 /**
  * Build the user code module. The code is wrapped in an exported async
