+++ executor/packages/sandbox-host/src/isolate/globals.isolate.js
@@ -1,0 +1,2 @@
+// Captured before user code module is evaluated (this module is imported first).
+export const ResponseJson = Response.json.bind(Response);
